+++ main_clean.py
@@ -1,19 +1,6 @@
-from __future__ import division
 import json
 import pandas as pd
-import numpy as np
-import requests
 import nltk
-import string
-import re
-import os
-from os import path
-from time import sleep
-from collections import Counter
-from nltk.tokenize import TweetTokenizer
-from nltk.classify import NaiveBayesClassifier
-from nltk.corpus import twitter_samples
-from random import shuffle
 
 def cleanTweet(tweet):
     # make lowercase
@@ -75,7 +62,3 @@ def main(twtInfo:object):
         ret[i]["text"] = clean_data_tweets[i]
         ret[i]["id"] = data_id[i]
     return pd.Series(ret).to_json(orient="records")
-
-dat = main("C:/Users/Matt/Documents/GitHub/twitter-sentiment-analysis/test.json")
-with open("test_clean_out.json", "w+") as out:
-    out.write(dat)
