+++ main_tokenize.py
@@ -1,19 +1,8 @@
-from __future__ import division
 import json
 import pandas as pd
-import numpy as np
-import requests
-import nltk
 import string
-import re
-import os
-from os import path
-from time import sleep
-from collections import Counter
+import nltk
 from nltk.tokenize import TweetTokenizer
-from nltk.classify import NaiveBayesClassifier
-from nltk.corpus import twitter_samples
-from random import shuffle
 
 def tokenizeTweets(tweetList):
     useless_ones = nltk.corpus.stopwords.words("english") + list(string.punctuation)
@@ -30,19 +19,6 @@ def tokenizeTweet(tweet):
     wordlist = [word for word in tokenizer.tokenize(tweet) if word not in useless_ones]
     return wordlist
 
-def stemTweets(tweetList):
-    sno = nltk.stem.SnowballStemmer("english")
-    retTweetList = []
-    for words in tweetList:
-        stemmed_words = [sno.stem(word) for word in words]
-        retTweetList.append(stemmed_words)
-    return retTweetList
-
-def stemTweet(tweet):
-    sno = nltk.stem.SnowballStemmer("english")
-    stemmed_words = [sno.stem(word) for word in tweet]
-    return stemmed_words
-
 def main(twtInfo:object):
     clean_data_tweets = pd.read_json(twtInfo, orient="records")
     nltk.download("stopwords")
@@ -52,16 +28,10 @@ def main(twtInfo:object):
     data_tc_tweets = []
     for tweet in tweets:
         data_tc_tweets.append(tokenizeTweet(tweet))
-    data_tcs_tweets = []
-    for tweet in data_tc_tweets:
-        data_tcs_tweets.append(stemTweet(tweet))
     ret = []
-    for i in range(len(data_tcs_tweets)):
+    for i in range(len(data_tc_tweets)):
         ret.append({})
-        ret[i]["text"] = data_tcs_tweets[i]
+        ret[i]["text"] = data_tc_tweets[i]
         ret[i]["id"] = data_id[i]
     return pd.Series(ret).to_json(orient="records")
 
-dat = main("C:/Users/Matt/Documents/GitHub/twitter-sentiment-analysis/test_clean_out.json")
-with open("test_token_out.json", "w+") as out:
-    out.write(dat)
