+++ test_lancaster_out.json
@@ -1,0 +1,1 @@
+[{"text":["followfriday","@france_inte","@pkuchly57","@milipol_paris","top","eng","memb","commun","week",":)"],"id":624495129303351296},{"text":["@lamb2ja","hey","jam","od",":\/","pleas","cal","contact","cent","02392441234","abl","assist",":)","many","thank"],"id":624495128510656513},{"text":["@despiteofficial","list","last","night",":)","blee","amaz","track","scotland"],"id":624495126535106560},{"text":["@97sides","congr",":)"],"id":624495126170238976},{"text":["yeaaaah","yippppy","accnt","ver","rqst","success","got","blu","tick","mark","fb","profil",":)","15","day"],"id":624495124148391936},{"text":["@bhaktisbanter","@pallaviruhail","on","irresist",":)","flipkartfashionfriday"],"id":624495123343118337},{"text":["lik","keep","lov","custom","wait","long","hop","enjoy","happy","friday","lwwf",":)"],"id":624495122982408192},{"text":["hopeless","tmr",":("],"id":624530164626534400},{"text":["everyth","kid","sect","ike","cut","sham","i'm","near","19","2","month",":("],"id":624530162890219521},{"text":["@hegelbon","heart","slid","wast","basket",":("],"id":624530162600812544},{"text":["@ketchburning","hat","japanes","cal","ban",":(",":("],"id":624530161011048448},{"text":["dang","start","next","week","work",":("],"id":624530159928934400},{"text":["oh","god","baby","fac",":("],"id":624530159593394176},{"text":["@rileymcdonough","mak","smil",":("],"id":624530150034571264}]
