+++ test_porter_out.json
@@ -1,0 +1,1 @@
+[{"text":["followfriday","@france_int","@pkuchly57","@milipol_pari","top","engag","member","commun","week",":)"],"id":624495129303351296},{"text":["@lamb2ja","hey","jame","odd",":\/","pleas","call","contact","centr","02392441234","abl","assist",":)","mani","thank"],"id":624495128510656513},{"text":["@despiteoffici","listen","last","night",":)","bleed","amaz","track","scotland"],"id":624495126535106560},{"text":["@97side","congrat",":)"],"id":624495126170238976},{"text":["yeaaaah","yippppi","accnt","verifi","rqst","succeed","got","blue","tick","mark","fb","profil",":)","15","day"],"id":624495124148391936},{"text":["@bhaktisbant","@pallaviruhail","one","irresist",":)","flipkartfashionfriday"],"id":624495123343118337},{"text":["like","keep","love","custom","wait","long","hope","enjoy","happi","friday","lwwf",":)"],"id":624495122982408192},{"text":["hopeless","tmr",":("],"id":624530164626534400},{"text":["everyth","kid","section","ikea","cute","shame","i'm","nearli","19","2","month",":("],"id":624530162890219521},{"text":["@hegelbon","heart","slide","wast","basket",":("],"id":624530162600812544},{"text":["@ketchburn","hate","japanes","call","bani",":(",":("],"id":624530161011048448},{"text":["dang","start","next","week","work",":("],"id":624530159928934400},{"text":["oh","god","babi","face",":("],"id":624530159593394176},{"text":["@rileymcdonough","make","smile",":("],"id":624530150034571264}]
