+++ test_token_out.json
@@ -1,1 +1,1 @@
-[{"text":["followfriday","@france_int","@pkuchly57","@milipol_pari","top","engag","member","communiti","week",":)"],"id":624495129303351296},{"text":["@lamb2ja","hey","jame","odd",":\/","pleas","call","contact","centr","02392441234","abl","assist",":)","mani","thank"],"id":624495128510656513},{"text":["@despiteoffici","listen","last","night",":)","bleed","amaz","track","scotland"],"id":624495126535106560},{"text":["@97side","congrat",":)"],"id":624495126170238976},{"text":["yeaaaah","yippppi","accnt","verifi","rqst","succeed","got","blue","tick","mark","fb","profil",":)","15","day"],"id":624495124148391936},{"text":["@bhaktisbant","@pallaviruhail","one","irresist",":)","flipkartfashionfriday"],"id":624495123343118337},{"text":["like","keep","love","custom","wait","long","hope","enjoy","happi","friday","lwwf",":)"],"id":624495122982408192},{"text":["hopeless","tmr",":("],"id":624530164626534400},{"text":["everyth","kid","section","ikea","cute","shame","i'm","near","19","2","month",":("],"id":624530162890219521},{"text":["@hegelbon","heart","slide","wast","basket",":("],"id":624530162600812544},{"text":["@ketchburn","hate","japanes","call","bani",":(",":("],"id":624530161011048448},{"text":["dang","start","next","week","work",":("],"id":624530159928934400},{"text":["oh","god","babi","face",":("],"id":624530159593394176},{"text":["@rileymcdonough","make","smile",":("],"id":624530150034571264}]
+[{"text":["followfriday","@france_inte","@pkuchly57","@milipol_paris","top","engaged","members","community","week",":)"],"id":624495129303351296},{"text":["@lamb2ja","hey","james","odd",":\/","please","call","contact","centre","02392441234","able","assist",":)","many","thanks"],"id":624495128510656513},{"text":["@despiteofficial","listen","last","night",":)","bleed","amazing","track","scotland"],"id":624495126535106560},{"text":["@97sides","congrats",":)"],"id":624495126170238976},{"text":["yeaaaah","yippppy","accnt","verified","rqst","succeed","got","blue","tick","mark","fb","profile",":)","15","days"],"id":624495124148391936},{"text":["@bhaktisbanter","@pallaviruhail","one","irresistible",":)","flipkartfashionfriday"],"id":624495123343118337},{"text":["like","keep","lovely","customers","waiting","long","hope","enjoy","happy","friday","lwwf",":)"],"id":624495122982408192},{"text":["hopeless","tmr",":("],"id":624530164626534400},{"text":["everything","kids","section","ikea","cute","shame","i'm","nearly","19","2","months",":("],"id":624530162890219521},{"text":["@hegelbon","heart","sliding","waste","basket",":("],"id":624530162600812544},{"text":["@ketchburning","hate","japanese","call","bani",":(",":("],"id":624530161011048448},{"text":["dang","starting","next","week","work",":("],"id":624530159928934400},{"text":["oh","god","babies","faces",":("],"id":624530159593394176},{"text":["@rileymcdonough","make","smile",":("],"id":624530150034571264}]
